requirements.txt
# Core data processing
pandas>=2.1.4
numpy>=1.26.4
pyarrow>=15.0.0

# Machine Learning & NLP
scikit-learn>=1.3.2

# Transformers and sentence-transformers
transformers>=4.35.2
sentence-transformers>=2.2.2

# Vector Store (CPU-only, compatible with Python 3.10+)
faiss-cpu>=1.7.4

# Model persistence
joblib>=1.3.2

# Web App
streamlit>=1.28.1

# Generative AI
openai>=1.3.9

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0

# Additional utilities
tqdm>=4.66.1
click>=8.1.7